fix(db): widen badges.score to integer and clear saturated points
POINTS_EARNED accumulates over a user's lifetime (-20 per skipped track),
which overflowed the smallint column (max ±32,767). On overflow Postgres
raised "smallint out of range", the error was swallowed by BadgeService's
try/catch, and the score silently froze at the floor — one imported legacy
user sat at -32,726.

- schema: badges.score smallint -> integer (0001)
- data fix: reset POINTS_EARNED rows saturated at the floor to 0 (0002)

Both migrations are idempotent.

+++ packages/db/schema.ts
@@ -187,7 +187,10 @@ export const badges = pgTable(
       .references(() => users.id, { onDelete: "cascade" }),
     lastUpdated: timestamp("last_updated").defaultNow().notNull(),
     name: badgeEnum("name").notNull(),
-    score: smallint("score").default(0).notNull(),
+    // `integer`, not `smallint`: POINTS_EARNED accumulates over a user's
+    // lifetime (e.g. -20 per skipped track) and can exceed smallint's
+    // ±32,767 range, overflowing and silently freezing the score.
+    score: integer("score").default(0).notNull(),
   },
   (t) => [unique().on(t.userId, t.name)],
 );
